backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
werkzeug==3.0.1
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
pillow==10.1.0
tensorflow==2.15.0
python-dotenv==1.0.0
gunicorn==21.2.0
pymysql==1.1.0
bcrypt==4.1.1
